split single file in multiple parts
