+++ .travis.yml
@@ -1,4 +1,4 @@
 language: go
 go:
-    - 1.1
-script: go test -v
+    - 1.4.2
+script: go test -v -cover
